requirements.txt
streamlit
pandas
numpy
python-dotenv
faiss-cpu
sentence-transformers
google-generativeai
